requirements.txt
pandas
scikit-learn
numpy
langchain
langchain-core
langchain-huggingface
langchain-chroma
langchain-groq
transformers
sentence-transformers
torch
huggingface-hub
langchain-experimental==0.0.62
langchain-qdrant==0.1.1
python-dotenv
qdrant-client==1.10.1
streamlit
pypdfium2==4.30.0
langchain-community==0.2.6
arxiv
wikipedia
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
pytube
unstructured
numexpr
ipykernel
jupyter
sentence_transformers
gradio
google-generativeai
langchain-google-genai
faiss-cpu
PyPDf2
ctransformers
chainlit
openai
langchain_nvidia_ai_endpoints
faiss-cpu